+++ Scripts/Controllers/DSC_Dialogue_GameObjectGroupController.cs
@@ -1,18 +1,17 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
+using DSC.Core;
 
 namespace DSC.DialogueSystem
 {
-    public class DSC_Dialogue_GameObjectGroupController : MonoBehaviour
+    public class DSC_Dialogue_GameObjectGroupController : BaseComponentGroupController
     {
         #region Variable
 
         #region Variable - Inspector
 #pragma warning disable 0649
 
-        [SerializeField] DSC_Dialogue_DataController m_hDataController;
-        [SerializeField] GameObject[] m_arrGameObject;
+        [SerializeField] protected DSC_Dialogue_DataController m_hDataController;
+        [SerializeField] protected GameObject[] m_arrGameObject;
 
 #pragma warning restore 0649
         #endregion
@@ -21,22 +20,22 @@ namespace DSC.DialogueSystem
 
         #region Base - Mono
 
-        private void Awake()
+        protected virtual void Awake()
         {
             if (m_hDataController && m_hDataController.DialogueEventDataList != null)
             {
-                m_hDataController.DialogueEventDataList.Add(new DialogueEventData_GameObject
+                m_hDataController.DialogueEventDataList.Add(new DialogueEventData_GroupController<DSC_Dialogue_GameObjectGroupController>
                 {
                     m_hGroupController = this,
                 });
             }
         }
 
-        private void OnDestroy()
+        protected virtual void OnDestroy()
         {
             if (m_hDataController != null && m_hDataController.DialogueEventDataList != null)
             {
-                m_hDataController.DialogueEventDataList.Remove<DialogueEventData_GameObject>();
+                m_hDataController.DialogueEventDataList.Remove<DialogueEventData_GroupController<DSC_Dialogue_GameObjectGroupController>>();
             }
         }
 
@@ -44,11 +43,20 @@ namespace DSC.DialogueSystem
 
         #region Main
 
-        public virtual void SetActiveGameObject(int nIndex,bool bActive)
+        public override void SetEnable(int nIndex, bool bEnable)
         {
             var hGameObject = GetGameObject(nIndex);
-            hGameObject?.SetActive(bActive);
+            hGameObject?.SetActive(bEnable);
         }
+
+        public override void SetAllEnable(bool bEnable)
+        {
+            for(int i = 0; i < m_arrGameObject.Length; i++)
+            {
+                m_arrGameObject[i]?.SetActive(bEnable);
+            }
+        }
+
 
         #endregion
 
@@ -62,6 +70,7 @@ namespace DSC.DialogueSystem
             return m_arrGameObject[nIndex];
         }
 
+       
         #endregion
     }
 }