+++ Scripts/DialogueDatas/Events/Datas/DialogueEventData_GroupController.cs
@@ -1,0 +1,9 @@
+﻿using DSC.Core;
+
+namespace DSC.DialogueSystem
+{
+    public struct DialogueEventData_GroupController<GroupController> : IDialogueEventData where GroupController : BaseComponentGroupController
+    {
+        public GroupController m_hGroupController;
+    }
+}