+++ Scripts/DialogueDatas/Events/DialogueEvent_ImageColor.cs
@@ -26,7 +26,7 @@ namespace DSC.DialogueSystem
         {
             base.OnStart(lstData);
 
-            if (!lstData.TryGetData(out DialogueEventData_Image hOutData))
+            if (!lstData.TryGetData(out DialogueEventData_GroupController<DSC_Dialogue_ImageGroupController> hOutData))
                 return;
 
             var hImageGroupController = hOutData.m_hGroupController;