+++ Scripts/DialogueDatas/Events/DialogueEvent_ImageSet.cs
@@ -45,7 +45,7 @@ namespace DSC.DialogueSystem
         {
             base.OnStart(lstData);
 
-            if (!lstData.TryGetData(out DialogueEventData_Image hOutData))
+            if (!lstData.TryGetData(out DialogueEventData_GroupController<DSC_Dialogue_ImageGroupController> hOutData))
                 return;
 
             var hImageGroupController = hOutData.m_hGroupController;
@@ -100,7 +100,7 @@ namespace DSC.DialogueSystem
         {
             DSC_Dialogue_ImageGroupController hResult = null;
 
-            if (lstData.TryGetData(out DialogueEventData_Image hOutData))
+            if (lstData.TryGetData(out DialogueEventData_GroupController<DSC_Dialogue_ImageGroupController> hOutData))
             {
                 hResult = hOutData.m_hGroupController;
             }