+++ Scripts/DialogueDatas/Events/DialogueEvent_RawImageColor.cs
@@ -26,7 +26,7 @@ namespace DSC.DialogueSystem
         {
             base.OnStart(lstData);
 
-            if (!lstData.TryGetData(out DialogueEventData_RawImage hOutData))
+            if (!lstData.TryGetData(out DialogueEventData_GroupController<DSC_Dialogue_RawImageGroupController> hOutData))
                 return;
 
             var hGroupController = hOutData.m_hGroupController;